halcmd: waitusr: avoid race condition
The sequence
    unloadusr foo
    waitusr foo
is totally sensible and is used in the tests.  However, there is a
race between sending the termination signal and waiting for the
component to exit.  Treat the component not existing at the start of
'waitusr' as success, even though it can also indicate a typo in the
component name.

It is believed that this will fix the occasional load-dependent
failure in the test 'userspace-count-names'.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>